backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorclass==2.2.0
distro
distro-info
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-sslserver==0.22
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
docopt==0.6.2
idna==2.10
langdetect==1.0.8
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
netifaces==0.10.9
packaging==20.4
Pillow==7.2.0
pip-review==1.1.0
pip-upgrader==1.4.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycairo==1.19.1
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-unixsocket==0.2.0
six==1.15.0
sqlparse==0.3.1
ssh-import-id==5.10
terminaltables==3.1.0
transliterate==1.10.2
urllib3
